Update project structure.
Move directory into sub-directory; create a run.py file for main entry point; move routes and models to individual directories and files.
