+++ flash_learning/__init__.py
@@ -1,0 +1,18 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+from flash_learning.config import Config
+
+db = SQLAlchemy()
+
+
+def create_app(config_class=Config):
+    app = Flask(__name__)
+    app.config.from_object(config_class)
+
+    db.init_app(app)
+
+    from flash_learning.main.routes import main
+    app.register_blueprint(main)
+
+    return app
