+++ flash_learning/config.py
@@ -1,0 +1,3 @@
+class Config:
+    SQLALCHEMY_DATABASE_URI = "sqlite:///site.db"
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
