+++ flash_learning/main/routes.py
@@ -1,0 +1,9 @@
+from flask import Blueprint, render_template
+
+
+main = Blueprint("main", __name__)
+
+
+@main.route('/', methods=["GET", "POST"])
+def index():
+    return render_template("index.html")
